+++ lib/src/datastore/options.dart
@@ -1,18 +1,18 @@
-
 /// Specifies read consistency.
 class ReadConsistency {
   /// Strong consistency.
-  static const strong = const ReadConsistency._("STRONG");
+  static const ReadConsistency strong = const ReadConsistency._("STRONG");
 
   /// Eventual consistency.
-  static const eventual = const ReadConsistency._("EVENTUAL");
+  static const ReadConsistency eventual = const ReadConsistency._("EVENTUAL");
 
   static const List<ReadConsistency> values = const [strong, eventual];
 
   /// Look up read consistency by name.
   static ReadConsistency lookup(String name) {
     if (_nameToValue == null) {
-      _nameToValue = new Map.fromIterable(values, key: (ReadConsistency value) => value.name);
+      _nameToValue =
+          new Map<String, ReadConsistency>.fromIterable(values, key: (value) => (value as ReadConsistency).name);
     }
     return _nameToValue[name] ?? new ReadConsistency._(name);
   }